optimization of views
