+++ src/main/java/org/ohdsi/webapi/common/generation/AnalysisGenerationInfo.java
@@ -1,38 +1,16 @@
 package org.ohdsi.webapi.common.generation;
 
-import org.ohdsi.webapi.shiro.Entities.UserEntity;
-
 import javax.persistence.Column;
 import javax.persistence.Embeddable;
-import javax.persistence.FetchType;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
 
 @Embeddable
-public class  AnalysisGenerationInfo {
+public class AnalysisGenerationInfo extends AnalysisGenerationBaseInfo{
 
     @Column(name = "design")
     protected String design;
 
-    @Column(name = "hash_code")
-    protected Integer hashCode;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "created_by_id", updatable = false)
-    protected UserEntity createdBy;
-
     public String getDesign() {
 
         return design;
-    }
-
-    public Integer getHashCode() {
-
-        return hashCode;
-    }
-
-    public UserEntity getCreatedBy() {
-
-        return createdBy;
     }
 }
