+++ src/main/java/org/ohdsi/webapi/pathway/PathwayService.java
@@ -4,6 +4,7 @@ import org.ohdsi.webapi.job.JobExecutionResource;
 import org.ohdsi.webapi.pathway.domain.PathwayAnalysisEntity;
 import org.ohdsi.webapi.pathway.domain.PathwayAnalysisGenerationEntity;
 import org.ohdsi.webapi.pathway.dto.internal.PathwayAnalysisResult;
+import org.ohdsi.webapi.shiro.annotations.PathwayAnalysisGenerationId;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 
@@ -43,4 +44,6 @@ public interface PathwayService {
     void cancelGeneration(Integer pathwayAnalysisId, Integer sourceId);
 
     int countLikeName(String copyName);
+
+    String findDesignByGenerationId(@PathwayAnalysisGenerationId final Long id);
 }
