+++ src/main/java/org/ohdsi/webapi/pathway/domain/PathwayAnalysisGenerationEntity.java
@@ -1,7 +1,6 @@
 package org.ohdsi.webapi.pathway.domain;
 
 import org.ohdsi.webapi.common.generation.CommonGeneration;
-import org.ohdsi.webapi.pathway.converter.SerializedPathwayAnalysisToPathwayAnalysisConverter;
 
 import javax.persistence.Entity;
 import javax.persistence.FetchType;
@@ -20,9 +19,5 @@ public class PathwayAnalysisGenerationEntity extends CommonGeneration {
     public PathwayAnalysisEntity getPathwayAnalysis() {
 
         return pathwayAnalysis;
-    }
-
-    public PathwayAnalysisEntity getDesign() {
-        return new SerializedPathwayAnalysisToPathwayAnalysisConverter().convertToEntityAttribute(this.info.getDesign());
     }
 }
